--- a/TimeTrace/src/components/LogTable.tsx
+++ b/TimeTrace/src/components/LogTable.tsx
@@ -1,7 +1,6 @@
 import { useState, useEffect, useContext } from "react";
 import { AppdataContext } from "../context/AppContext";
 import { FileLine, mapEventsToFileLine } from '../models/Types/FileLine';
-import { log } from "console";
 
 type LogTableProps = {
     mappingsAreEditable: boolean;