… README.md
… TimeTrace/src/models/MonaaZone.ts
… TimeTrace/src/models/QueryHandler.ts
… TimeTrace/src/pages/LogPage.tsx
… TimeTrace/src/tests/models/LogHandler.test.ts
… TimeTrace/src/tests/models/LogSearcher.test.ts
… experiments/ConsoleTimeFormatter/performance_formatter.py