## Used log:
log1Col_20000.txt

### Search:  ```ab$```

### Time of running findZones method in LogSearcher with binary search with TRE: ```ab$```
| Run Time (ms) |
|---------------|
| 7.367919921875 |
| 6.8291015625 |
| 6.880859375 |
| 6.157958984375 |
| 8.06689453125 |
| 6.14794921875 |
| 6.091064453125 |
| 6.205810546875 |
| 6.443115234375 |
| 6.302001953125 |
| 7.56591796875 |
| 6.35205078125 |
| 7.14794921875 |
| 6.8349609375 |
| 6.575927734375 |
| 6.013916015625 |
| 7.9169921875 |
| 6.589111328125 |
| 7.76416015625 |
| 5.944091796875 |
| 6.39892578125 |
| 6.882080078125 |
| 7.30908203125 |
| 6.196044921875 |
| 6.0498046875 |
| 6.122802734375 |
| 6.59716796875 |
| 6.168212890625 |
| 6.283203125 |
| 6.22607421875 |
| 5.767822265625 |
| 6.080078125 |
| 6.3759765625 |
| 9.52294921875 |
| 7.421142578125 |
| 7.280029296875 |
| 6.767822265625 |
| 6.737060546875 |
| 7.554931640625 |
| 6.14599609375 |
| 6.02978515625 |
| 10.028076171875 |
| 6.14501953125 |
| 6.326904296875 |
| 6.966064453125 |
| 7.6669921875 |
| 6.880859375 |
| 6.947998046875 |
| 7.1240234375 |
| 6.881103515625 |
| 7.458984375 |
| 7.68017578125 |
| 7.2060546875 |
| 6.158203125 |
| 8.10791015625 |
| 7.009765625 |
| 6.971923828125 |
| 8.9921875 |
| 6.871826171875 |

Average Run Time: 6.8904884467690675 ms


### Time of running findZones method in LogSearcher with TRE: ```ab$```
| Time | Without optimization | With optimization W/O binary search|
| ----- | ------------- | ------------- | 
| ----- | 2289.97802734375  | 11.23193359375 | 
| ----- | 2496.590087890625  |  11.529052734375 | 
| ----- | 2711.100830078125 | 12.9208984375 | 
| ----- | 3092.488037109375  | 11.0419921875 | 
| ----- | 2326.285888671875  | 9.10302734375 | 
| ----- | 2299.099853515625  | 7.135009765625 | 
| ----- | 3563.39501953125 | 7.58203125 | 
| ----- | 2285.85693359375 | 8.15087890625 | 
| ----- | 2265.970947265625  | 7.572998046875 | 
| ----- | 2320.158935546875  | 7.428955078125 | 
| ----- | 2296.489013671875 | 8.05712890625 | 


## Optimized code:
```ts
import { extractTimeStamp } from "./helpers/extractTimeStamp";
import { MonaaZone } from "./MonaaZone";
import { SearchInterval } from "./SearchInterval";

export class LogSearcher {
  findZones(logFile: string[], SearchIntervals: SearchInterval[]): MonaaZone[] {
    console.time("findZones");
    const MonaaZoneMatches: MonaaZone[] = [];
    const [timestamps, averageTimegrowth] = this.getTimestampInfo(logFile);

    for (let i = 0; i < SearchIntervals.length; i++) {
      let foundmatch = new MonaaZone();
      let startingIndex = this.findStartingIndex(
        timestamps,
        SearchIntervals[i],
        averageTimegrowth
      );

      for (let j = startingIndex; j < timestamps.length; j++) {
        const timestamp = timestamps[j];

        if (
          timestamp >= SearchIntervals[i].start &&
          timestamp <= SearchIntervals[i].end
        ) {
          foundmatch.lineMatches.push(j);
        } else if (timestamp > SearchIntervals[i].end) {
          break;
        }
      }
      MonaaZoneMatches.push(foundmatch);
    }
    console.timeEnd("findZones");
    return MonaaZoneMatches;
  }

  getTimestampInfo(logFile: string[]): [number[], number] {
    let prevLineTime: number = 0;
    let averageTimegrowth: number = 0;
    const timestamps: number[] = [];

    logFile.forEach((line: string) => {
      const timestampISO: string = extractTimeStamp(line);
      const eventTimeStamp = new Date(timestampISO).getTime();
      timestamps.push(eventTimeStamp);
      averageTimegrowth =
        (averageTimegrowth + (eventTimeStamp - prevLineTime)) / 2;
      prevLineTime = eventTimeStamp;
    });

    return [timestamps, averageTimegrowth];
  }

  findStartingIndex(
    timestamps: number[],
    searchInterval: SearchInterval,
    averageTimegrowth: number
  ): number {
    const firstTimestamp = timestamps[0];
    const difference = searchInterval.start - firstTimestamp;
    const multiplum = difference / averageTimegrowth;
    let startingIndex = Math.floor(multiplum);

    while (timestamps[startingIndex] > searchInterval.start) {
      startingIndex--;
    }

    // Using binary search to find the correct starting index
    startingIndex = this.binarySearch(
      timestamps,
      searchInterval.start,
      startingIndex
    );

    return startingIndex;
  }

  // Binary search function to find the first index in the log to search from
  binarySearch(
    timestamps: number[],
    target: number,
    startingIndex: number
  ): number {
    let left = startingIndex;
    let right = timestamps.length - 1;
    let resultIndex = -1;

    while (left <= right) {
      const mid = Math.floor((left + right) / 2);
      if (timestamps[mid] >= target) {
        resultIndex = mid; // Update result index, as we found an element greater than or equal to the target
        right = mid - 1; // Continue searching in the left half
      } else {
        left = mid + 1; // Continue searching in the right half
      }
    }

    return resultIndex;
  }
}

```


## Unoptimized code:
```ts
import { MonaaZone } from "./MonaaZone";
import { SearchInterval } from "./SearchInterval";
import { extractTimeStamp } from "./helpers/extractTimeStamp";

export class LogSearcher {
  findZones(logFile: string[], SearchIntervals: SearchInterval[]): MonaaZone[] {
    console.time("findZones");
    const MonaaZoneMatches: MonaaZone[] = [];

    for (let i = 0; i < SearchIntervals.length; i++) {
      let foundmatch = new MonaaZone();
      for (let j = 0; j < logFile.length; j++) {
        const element = logFile[j];
        let timestampISO: string = extractTimeStamp(element);
        const eventTimeStamp = new Date(timestampISO).getTime();
        if (
          eventTimeStamp >= SearchIntervals[i].start &&
          eventTimeStamp <= SearchIntervals[i].end
        ) {
          foundmatch.lineMatches.push(j);
        } else if (eventTimeStamp > SearchIntervals[i].end) {
          break;
        }
      }

      MonaaZoneMatches.push(foundmatch);
    }
    console.timeEnd("findZones");
    console.log("MonaaZoneMatches", MonaaZoneMatches);
    return MonaaZoneMatches;
  }
}
```